Codecs: Decode E-AC-3 spectral extension
The block parser reads the spectral extension strategy and coordinates
(Table E1.4) instead of refusing them: the frame starts from Table E2.11's
banding, coordinates are mandatory the first time a channel extends in a
frame, and blend factors are recomputed when they arrive. Per §E3.3,
cplendf is derived from spxbegf and can go to -2, so it and
ac3CouplingEndMantissa are now signed; an extended channel sends no chbwcod
and ends where extension begins; and rematrixing without coupling uses 3 or
4 bands by spxbegf. Synthesis runs after rematrixing, the only order in which
it copies a channel's own finished low band -- Annex E does not state the
place explicitly.

The noise generators now outlive the per-frame state reset. AC3Dither lived
in AC3FrameState, which is cleared at every syncframe, so the dither
sequence restarted 31 times a second despite its comment saying otherwise;
it and the new spectral extension noise are now carried across.

+++ include/codecs/ac3/AC3Exponents.h
@@ -77,7 +77,7 @@ unsigned ac3ChannelEndMantissa(uint8_t chbwcod);
 
 /// Coupling channel bin range, A/52 §7.1.3.
 unsigned ac3CouplingStartMantissa(uint8_t cplbegf);
-unsigned ac3CouplingEndMantissa(uint8_t cplendf);
+unsigned ac3CouplingEndMantissa(int cplendf);
 
 /// The LFE channel is fixed: bins 0 to 7, always two groups, always D15.
 constexpr unsigned kLfeEndMantissa = 7;
